Update react-ui-kit dependency version (#159)
* Update react-ui-kit dependency version

* Update react-ui-kit dependency version
